requirements.txt
streamlit
PyMuPDF
numpy
python-dotenv
groq
sentence-transformers